# openapi-postman.yaml
openapi: 3.0.0
info:
  title: Inbound Management API
  version: 1.0.0
  description: >
    Dokumentasi API untuk Inbound Management System menggunakan Next.js dan Prisma.
    Seluruh endpoint secara global membutuhkan autentikasi menggunakan Bearer Token (JWT),
    kecuali endpoint login.
servers:
  - url: http://localhost:3000
    description: Development Server
security:
  - bearerAuth: []
paths:
...
  /api/manifests/submit:
    post:
      summary: Submit Manifest
      description: Menyimpan dokumen manifes dan langsung dikunci (status LOCKED). Meng-generate QR Code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManifestCreateInput'
      responses:
        '201':
          description: Manifest berhasil disubmit dan dilock
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Manifest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Manifest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        manifest_number:
          type: string
        vendor_id:
          type: string
          format: uuid
        driver_name:
          type: string
        vehicle_plate:
          type: string
        departure_date:
          type: string
          format: date-time
        estimated_arrival:
          type: string
          format: date-time
        is_locked:
          type: boolean
        locked_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [DRAFT, LOCKED, CHECKING, COMPLETED, DISCREPANCY]
        created_at:
          type: string
          format: date-time
    ManifestCreateInput:
      type: object
      properties:
        vendor_id:
          type: string
          format: uuid
        driver_name:
          type: string
        vehicle_plate:
          type: string
        estimated_arrival:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: string
                format: uuid
              expected_qty:
                type: integer
              expected_boxes:
                type: integer
              batch_code:
                type: string
...